# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Barefoot Nomad",
    "description" : "Make company global travel and accommodation easy and convenient for the strong workforce of savvy members of staff, by leveraging the modern web."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/",
  "securityDefinitions": {},
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admin/users": {
      "put": {
        "description": "rchange a new role",
        "summary": "change role",
        "tags": [
          "Super Admin"
        ],
        "operationId": "ChangeRole",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    
    }
  }
}